# Increase riscv implementation in DepthwiseConvKernel