@wf1594149160 2024-06-06
This is our attempt at implementing RVV. Based on the code implementation, it does not negatively impact the original PyTorch build. Additionally, we observed significant performance improvements when performing inference on MobileNet_v2. With input data of [1, 3, 224, 224], we saw a 3x performance increase. 
Has anyone followed up on this PR?
@mikaylagawarecki @ezyang