@ezyang 2024-06-10
This would be the very first riscv kernel we'd add to PyTorch. I'm not sure if we want to commit to maintaining all of these kernels. @malfet for more comments.